pyproject.toml
[project]
name = "briefy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "agno>=2.5.17",
    "apscheduler>=3.11.2",
    "discord-py>=2.7.1",
    "python-dotenv>=1.2.2",
]
